docs + server.go split
